RSpec: split up rmagick/draw_spec.rb
This is the first of many commits that will break up specs into one file
per method. It's a little unconventional to structure specs in this way
in Ruby, but here are a few benefits:

- Spec files are smaller, which makes them easier to navigate when
  looking for details relevant to what you're trying to test. This is
  especially valuable for classes that have dozens of methods where the
  test files can grow to several thousands of lines long.
- Easier to find specs for the method you need to test. You can
  now use your fuzzy finder to locate more specifically where the tests
  are.
- Encourages thinking about tests in a structured, per-method way.
  Structuring spec files this way is constraining in a way that pushes
  developers to think about which method is most relevant to what
  they're testing. It also saves some mental energy as developers don't
  have to spend as much time considering where tests belong in a much
  larger spec file, or what the description should be.
