+++ .rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2019-11-22 15:13:23 -0800 using RuboCop version 0.64.0.
+# on 2019-12-15 08:35:52 -0800 using RuboCop version 0.64.0.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -132,11 +132,11 @@ Lint/Void:
 Metrics/AbcSize:
   Max: 559
 
-# Offense count: 77
+# Offense count: 76
 # Configuration parameters: CountComments, ExcludedMethods.
 # ExcludedMethods: refine
 Metrics/BlockLength:
-  Max: 1381
+  Max: 1371
 
 # Offense count: 1
 # Configuration parameters: CountBlocks.
@@ -146,7 +146,7 @@ Metrics/BlockNesting:
 # Offense count: 9
 # Configuration parameters: CountComments.
 Metrics/ClassLength:
-  Max: 467
+  Max: 465
 
 # Offense count: 21
 Metrics/CyclomaticComplexity:
@@ -256,21 +256,19 @@ Performance/RedundantMatch:
     - 'examples/demo.rb'
     - 'spec/image_2_spec.rb'
 
-# Offense count: 3
+# Offense count: 2
 # Cop supports --auto-correct.
 RSpec/AlignLeftLetBrace:
   Exclude:
     - 'spec/rmagick/image/constitute_spec.rb'
     - 'spec/rmagick/image/from_blob_spec.rb'
-    - 'spec/rmagick/image/properties_spec.rb'
 
-# Offense count: 3
+# Offense count: 2
 # Cop supports --auto-correct.
 RSpec/AlignRightLetBrace:
   Exclude:
     - 'spec/rmagick/image/constitute_spec.rb'
     - 'spec/rmagick/image/from_blob_spec.rb'
-    - 'spec/rmagick/image/properties_spec.rb'
 
 # Offense count: 517
 # Cop supports --auto-correct.
@@ -279,7 +277,7 @@ RSpec/AlignRightLetBrace:
 RSpec/DescribedClass:
   Enabled: false
 
-# Offense count: 361
+# Offense count: 359
 # Configuration parameters: Max.
 RSpec/ExampleLength:
   Enabled: false
@@ -302,7 +300,7 @@ RSpec/HooksBeforeExamples:
     - 'spec/image_2_spec.rb'
     - 'spec/magick_2_spec.rb'
 
-# Offense count: 2896
+# Offense count: 2876
 # Configuration parameters: AssignmentOnly.
 RSpec/InstanceVariable:
   Exclude:
@@ -331,7 +329,7 @@ RSpec/MultipleDescribes:
   Exclude:
     - 'spec/fill_spec.rb'
 
-# Offense count: 552
+# Offense count: 553
 # Configuration parameters: AggregateFailuresByDefault.
 RSpec/MultipleExpectations:
   Max: 60
@@ -339,7 +337,7 @@ RSpec/MultipleExpectations:
 # Offense count: 3
 RSpec/Pending:
   Exclude:
-    - 'spec/rmagick/draw_spec.rb'
+    - 'spec/rmagick/draw/marshal_dump_spec.rb'
     - 'spec/rmagick/image/read_spec.rb'
 
 # Offense count: 2
@@ -380,7 +378,7 @@ Security/MarshalLoad:
     - 'lib/rmagick_internal.rb'
     - 'spec/image_2_spec.rb'
     - 'spec/image_list_2_spec.rb'
-    - 'spec/rmagick/draw_spec.rb'
+    - 'spec/rmagick/draw/marshal_dump_spec.rb'
 
 # Offense count: 2
 Style/AutoResourceCleanup:
@@ -394,7 +392,7 @@ Style/CaseEquality:
     - 'spec/enum_spec.rb'
     - 'spec/pixel_spec.rb'
 
-# Offense count: 6
+# Offense count: 5
 # Cop supports --auto-correct.
 # Configuration parameters: AutoCorrect, EnforcedStyle.
 # SupportedStyles: nested, compact
@@ -403,7 +401,6 @@ Style/ClassAndModuleChildren:
     - 'doc/ex/composite_layers.rb'
     - 'lib/rvg/to_c.rb'
     - 'spec/magick_spec.rb'
-    - 'spec/rmagick/draw_spec.rb'
 
 # Offense count: 24
 # Cop supports --auto-correct.
@@ -422,7 +419,7 @@ Style/CollectionMethods:
 Style/CommentedKeyword:
   Enabled: false
 
-# Offense count: 264
+# Offense count: 275
 # Cop supports --auto-correct.
 # Configuration parameters: Notice, AutocorrectNotice.
 Style/Copyright:
@@ -467,7 +464,7 @@ Style/FormatStringToken:
     - 'lib/rmagick_internal.rb'
     - 'lib/rvg/pathdata.rb'
 
-# Offense count: 264
+# Offense count: 275
 # Cop supports --auto-correct.
 # Configuration parameters: EnforcedStyle.
 # SupportedStyles: when_needed, always, never
@@ -496,7 +493,7 @@ Style/IfUnlessModifier:
 Style/InlineComment:
   Enabled: false
 
-# Offense count: 4451
+# Offense count: 4458
 # Cop supports --auto-correct.
 # Configuration parameters: IgnoreMacros, IgnoredMethods, AllowParenthesesInMultilineCall, AllowParenthesesInChaining, AllowParenthesesInCamelCaseMethod, EnforcedStyle.
 # SupportedStyles: require_parentheses, omit_parentheses
