+++ spec/rmagick/draw/kerning_writer_spec.rb
@@ -1,0 +1,7 @@
+RSpec.describe Magick::Draw, '#kerning=' do
+  let(:draw) { described_class.new }
+
+  it 'assigns without raising an error' do
+    expect { draw.kerning = 1 }.not_to raise_error
+  end
+end
